Reorganized scripts
