+++ eptobrick/extractor/objects/ahus.py
@@ -1,159 +1,117 @@
-class AHU():
-    pass
-class AHUComponent():
+from . import eplusbindings as epb
+from ..utils import utils as ut
 
-    def __init__(self, idfobj, objtype) -> None:
-        
-        #define base properties here
-        self.idfElement = idfobj
-        self.name = idfobj.name
-        self.type = objtype
-        self.air_inlets = None
-        self.air_outlets = None
-        self.data = {}
-        self.btype = None
+m = ut.Nuncius(debug=3)
 
-    def add_property(self, property, values):
-        self.data[property] = values
+def untangleAirLoopZones(idf):
+    zoneequipment = []
+    for zone in idf.ZoneHVAC_EquipmentConnections:
+        inlets = []
+        outlets = []
+        for nodelist in idf.NodeList:
+            if nodelist.name == zone.zone_air_inlet_node_or_nodelist_name:
+                inlets.extend(ut.getListComponents(nodelist, 'node', 'name'))
+            elif nodelist.name == zone.zone_return_air_node_or_nodelist_name or nodelist.name == zone.zone_air_exhaust_node_or_nodelist_name:
+                outlets.extend(ut.getListComponents(nodelist, 'node', 'name'))
+        # If the 'node or nodelist' did not refer to a nodelist, it means it is a single node.
+        if inlets == []:
+            inlets.append(zone.zone_air_inlet_node_or_nodelist_name)
+        if outlets == []:
+            if zone.zone_return_air_node_or_nodelist_name is not None:
+                outlets.append(zone.zone_return_air_node_or_nodelist_name)
+            if zone.zone_air_exhaust_node_or_nodelist_name is not None:
+                outlets.append(zone.zone_air_exhaust_node_or_nodelist_name)
+        curzone = [epb.ThermalZone(zone, ut.getEPType(zone), inlets, outlets).create()]
+        adus = ut.getListComponents(zone.zone_conditioning_equipment_list_name, 'zone_equipment', 'name')
+        for adu in adus:
+            terminalunit = adu.air_terminal_name
+            curzone.append(epb.AHUComponent(terminalunit, ut.getEPType(terminalunit)).create())
+        m.printMessage(f'Extracted zone and ADU nodes for {zone.zone_name.name}', lvl='debug')
+        m.printMessage(f'\t{zone.zone_name.name} outlets: {curzone[0].air_outlets}', lvl='debug')
+        m.printMessage(f'\t{terminalunit.name} inlets: {curzone[-1].air_inlets}', lvl='debug')
 
-    def add_inlet(self, name):
-        self.add_airconn('air_inlets', name)
 
-    def add_outlet(self, name):
-        self.add_airconn('air_outlets', name)
-
-    def add_airconn(self, conn, name):
-        if getattr(self, conn) is None:
-            setattr(self, conn, [name])
-        else:
-            getattr(self, conn).append(name)
-
-    def getListComponents(self, idflist, type, value):
-        components = []
-        i = 1
-        while True:
-            try:
-                components.append(getattr(idflist, f'{type}_{i}_{value}'))
-                i += 1
-            except:
-                break
-        return list(filter(None, components))
-
-    def create(self):
-        subtype = self.type.split(':')[0]
-        subsubtype = self.type.split(':')[1]
-        # Find type, generally speaking (e.g. fan, coil, terminal unit, etc.)
-        if subtype == 'Fan':
-            self.addFan()
-        elif subtype == 'OutdoorAir':
-            self.addOAS()
-        elif subtype == 'AirTerminal':
-            self.addTerminal()
-        elif subtype == 'Coil':
-            self.addCoil()
-        elif subtype == 'AirLoopHVAC':
-            if subsubtype == 'ZoneSplitter':
-                self.addSplitter()
-            elif subsubtype == 'ZoneMixer':
-                self.addMixer()
-        #elif subtype == 'AirLoopHVAC':
-        #    if subsubtype == 'UnitaryHeatPump':
-        #        self.addUHP()
-        return self
-
-    #### UNITARY SYSTEMS ####
-
-    def addUHP():
-        pass
-
-    #### FANS ####
-
-    def addFan(self):
-        self.btype = 'Fan'
-        self.add_property('fanEfficiency', float(self.idfElement.fan_total_efficiency))
-        self.add_property('motorEfficiency', float(self.idfElement.motor_efficiency))
-        self.add_property('cooledByFluid', bool(float(self.idfElement.motor_in_airstream_fraction)))
-        self.add_inlet(self.idfElement.air_inlet_node_name)
-        self.add_outlet(self.idfElement.air_outlet_node_name)
-        return self
-        #if self.type.split[1] == 'OnOff':
-        #    modelicabindings.FanOnOff()
-
-    #### COILS ####
-
-    def addCoil(self):
-        self.add_inlet(self.idfElement.air_inlet_node_name)
-        self.add_outlet(self.idfElement.air_outlet_node_name)
-
-        if self.type.split(':')[1] == 'Heating':
-            self.addHeatingCoil()
-        elif self.type.split(':')[1] == 'Cooling':
-            self.addCoolingCoil()
     
-    def addHeatingCoil(self):
-        self.btype = 'Heating_Coil'
-    def addCoolingCoil(self):
-        self.btype = 'Cooling_Coil'
+    return zoneequipment
 
-    #### OA ####
+def untangleAirLoopSupply(idfahu, idf):
+    supply_inlet = idfahu.supply_side_inlet_node_name
+    supply_outlet = idfahu.supply_side_outlet_node_names
+    demand_inlet = idfahu.demand_side_inlet_node_names
+    demand_outlet = idfahu.demand_side_outlet_node_name
 
-    def addOAS(self):
-        if self.type.split(':')[1] == 'Mixer':
-            self.btype = None
-            self.add_inlet(self.idfElement.outdoor_air_stream_node_name)
-            self.add_inlet(self.idfElement.return_air_stream_node_name)
-            self.add_outlet(self.idfElement.mixed_air_node_name)
-            self.add_outlet(self.idfElement.relief_air_stream_node_name)
+    conns = {
+        'sin' : supply_inlet,
+        'sout' : supply_outlet,
+        'din' : demand_inlet,
+        'dout' : demand_outlet
+    }
 
-    #### TERMINALS ####
+    ## SUPPLY ##
+    branches = ut.getListComponents(idfahu.branch_list_name, 'branch', 'name')
+    components = []
+    filtered = []
+    for branch in branches:
+        components.append(ut.getListComponents(branch, 'component', 'name'))
+        for componentlist in components:
+            for component in componentlist:
+                edgecases = componentEdgeCases(component)
+                if edgecases is not None:
+                    filtered.extend(edgecases)
+                else:
+                    filtered.append(epb.AHUComponent(component, ut.getEPType(component)).create())
+                ### This is where we check for component type! There is no further nesting.
+    
+    m.printMessage(f'Extracted supply branch for {idfahu.name}:', lvl='debug')
+    m.printMessage(f'\tObjects: {[comp.type for comp in filtered]}', lvl='debug')
+    m.printMessage(f'\tSupply inlets: {filtered[0].air_inlets}', lvl='debug')
+    m.printMessage(f'\tSupply outlets: {filtered[-1].air_outlets}', lvl='debug')
 
-    def addTerminal(self):
-        self.add_inlet(self.idfElement.air_inlet_node_name)
-        self.add_outlet(self.idfElement.air_outlet_node_name)
-        if self.type.split(':')[1] == 'SingleDuct':
-            self.singleDuctTerminal()
+    ## DEMAND ##
 
-    def singleDuctTerminal(self):
+    for supplypath in idf.AirLoopHVAC_SupplyPath:
+        if supplypath.supply_air_path_inlet_node_name == demand_inlet:
+            splitters = ut.getListComponents(supplypath, 'component', 'name')
+            for splitter in splitters:
+                filtered.append(epb.AHUComponent(splitter, ut.getEPType(splitter)).create())
 
-        if self.type.split(':')[2] == 'ConstantVolume':
-            self.CVTerminal()
+    m.printMessage(f'Extracted supply path for {idfahu.name}', lvl='debug')
+    m.printMessage(f'\tSupply path starts at: {filtered[-1].air_inlets}', lvl='debug')
+    m.printMessage(f'\tSupply path ends at: {filtered[-1].air_outlets}', lvl='debug')
+            
+    for returnpath in idf.AirLoopHVAC_ReturnPath:
 
-    def CVTerminal(self):
+        if returnpath.return_air_path_outlet_node_name == demand_outlet:
+            mixers = ut.getListComponents(returnpath, 'component', 'name')
+            for mixer in mixers:
+                filtered.append(epb.AHUComponent(mixer, ut.getEPType(mixer)).create())
 
-        if self.type.split(':')[1] == 'NoReheat':
-            self.btype = 'CAV'
-            #modelicabindings.CVNoReheat()
+    m.printMessage(f'Extracted return path for {idfahu.name}', lvl='debug')
+    m.printMessage(f'\tReturn path starts at: {filtered[-1].air_inlets}', lvl='debug')
+    m.printMessage(f'\tReturn path ends at: {filtered[-1].air_outlets}', lvl='debug')
 
-    #### OTHER ZONE EQUIPMENT ####
+    return filtered, conns
 
-    def addSplitter(self):
-        self.btype = None
-        for outlet in self.getListComponents(self.idfElement, 'outlet', 'node_name'):
-            self.add_outlet(outlet)
-        self.add_inlet(self.idfElement.inlet_node_name)
-
-    def addMixer(self):
-        self.btype = None
-        for inlet in self.getListComponents(self.idfElement, 'inlet', 'node_name'):
-            self.add_inlet(inlet)
-        self.add_outlet(self.idfElement.outlet_node_name)
-
-class ThermalZone(AHUComponent):
-
-    def __init__(self, idfobj, objtype, inlets, outlets) -> None:
-        self.idfElement = idfobj
-        self.name = idfobj.zone_name
-        self.type = objtype
-        self.air_inlets = inlets
-        self.air_outlets = outlets
-        self.data = {}
-        self.btype = None
-
-    def create(self):
-        self.btype = 'HVAC_Zone'
-        return self
-
-
+def componentEdgeCases(idfobj):
+    filtered = []
+    gettype = ut.getEPType(idfobj).split(':')
+    if ut.getEPType(idfobj) == 'AirLoopHVAC:OutdoorAirSystem':
+        oacomponents = ut.getListComponents(idfobj.outdoor_air_equipment_list_name, 'component', 'name')
+        for oacomponent in oacomponents:
+            filtered.append(epb.AHUComponent(oacomponent, ut.getEPType(oacomponent)).create())
+        return filtered
+    elif ut.getEPType(idfobj).split(':')[1] == 'UnitaryHeatPump':
+        hpcomponents = [
+            idfobj.supply_air_fan_name,
+            idfobj.heating_coil_name,
+            idfobj.cooling_coil_name,
+            idfobj.supplemental_heating_coil_name
+        ]
+        for hpcomponent in hpcomponents:
+            filtered.append(epb.AHUComponent(hpcomponent, ut.getEPType(hpcomponent)).create())
+        return filtered
+    else:
+        return None 
 
 ######## DANGER ZONE ######
 class FluidConnectionMap():
@@ -171,7 +129,7 @@ class FluidConnectionMap():
         connections = []
         if self.map[start][how] is None:
             for key in self.map.keys():
-                if how is 'downstream':
+                if how == 'downstream':
                     if any(self.map[key]['in']) in self.map[start]['out']:
                         connections.append(key)
                 else:
