+++ eptobrick/extractor/objects/eplusbindings.py
@@ -1,0 +1,143 @@
+from ..utils import utils as ut
+
+class AHUComponent():
+
+    def __init__(self, idfobj, objtype) -> None:
+        
+        #define base properties here
+        self.idfElement = idfobj
+        self.name = idfobj.name
+        self.type = objtype
+        self.air_inlets = None
+        self.air_outlets = None
+        self.data = {}
+        self.btype = None
+
+    def add_property(self, property, values):
+        self.data[property] = values
+
+    def add_inlet(self, name):
+        self.add_airconn('air_inlets', name)
+
+    def add_outlet(self, name):
+        self.add_airconn('air_outlets', name)
+
+    def add_airconn(self, conn, name):
+        if getattr(self, conn) is None:
+            setattr(self, conn, [name])
+        else:
+            getattr(self, conn).append(name)
+
+    def create(self):
+        subtype = self.type.split(':')[0]
+        subsubtype = self.type.split(':')[1]
+        # Find type, generally speaking (e.g. fan, coil, terminal unit, etc.)
+        if subtype == 'Fan':
+            self.addFan()
+        elif subtype == 'OutdoorAir':
+            self.addOAS()
+        elif subtype == 'AirTerminal':
+            self.addTerminal()
+        elif subtype == 'Coil':
+            self.addCoil()
+        elif subtype == 'AirLoopHVAC':
+            if subsubtype == 'ZoneSplitter':
+                self.addSplitter()
+            elif subsubtype == 'ZoneMixer':
+                self.addMixer()
+        #elif subtype == 'AirLoopHVAC':
+        #    if subsubtype == 'UnitaryHeatPump':
+        #        self.addUHP()
+        return self
+
+    #### UNITARY SYSTEMS ####
+
+    def addUHP():
+        pass
+
+    #### FANS ####
+
+    def addFan(self):
+        self.btype = 'Fan'
+        self.add_property('fanEfficiency', float(self.idfElement.fan_total_efficiency))
+        self.add_property('motorEfficiency', float(self.idfElement.motor_efficiency))
+        self.add_property('cooledByFluid', bool(float(self.idfElement.motor_in_airstream_fraction)))
+        self.add_inlet(self.idfElement.air_inlet_node_name)
+        self.add_outlet(self.idfElement.air_outlet_node_name)
+        return self
+        #if self.type.split[1] == 'OnOff':
+        #    modelicabindings.FanOnOff()
+
+    #### COILS ####
+
+    def addCoil(self):
+        self.add_inlet(self.idfElement.air_inlet_node_name)
+        self.add_outlet(self.idfElement.air_outlet_node_name)
+
+        if self.type.split(':')[1] == 'Heating':
+            self.addHeatingCoil()
+        elif self.type.split(':')[1] == 'Cooling':
+            self.addCoolingCoil()
+    
+    def addHeatingCoil(self):
+        self.btype = 'Heating_Coil'
+    def addCoolingCoil(self):
+        self.btype = 'Cooling_Coil'
+
+    #### OA ####
+
+    def addOAS(self):
+        if self.type.split(':')[1] == 'Mixer':
+            self.btype = None
+            self.add_inlet(self.idfElement.outdoor_air_stream_node_name)
+            self.add_inlet(self.idfElement.return_air_stream_node_name)
+            self.add_outlet(self.idfElement.mixed_air_node_name)
+            self.add_outlet(self.idfElement.relief_air_stream_node_name)
+
+    #### TERMINALS ####
+
+    def addTerminal(self):
+        self.add_inlet(self.idfElement.air_inlet_node_name)
+        self.add_outlet(self.idfElement.air_outlet_node_name)
+        if self.type.split(':')[1] == 'SingleDuct':
+            self.singleDuctTerminal()
+
+    def singleDuctTerminal(self):
+
+        if self.type.split(':')[2] == 'ConstantVolume':
+            self.CVTerminal()
+
+    def CVTerminal(self):
+
+        if self.type.split(':')[1] == 'NoReheat':
+            self.btype = 'CAV'
+            #modelicabindings.CVNoReheat()
+
+    #### OTHER ZONE EQUIPMENT ####
+
+    def addSplitter(self):
+        self.btype = None
+        for outlet in ut.getListComponents(self.idfElement, 'outlet', 'node_name'):
+            self.add_outlet(outlet)
+        self.add_inlet(self.idfElement.inlet_node_name)
+
+    def addMixer(self):
+        self.btype = None
+        for inlet in ut.getListComponents(self.idfElement, 'inlet', 'node_name'):
+            self.add_inlet(inlet)
+        self.add_outlet(self.idfElement.outlet_node_name)
+
+class ThermalZone(AHUComponent):
+
+    def __init__(self, idfobj, objtype, inlets, outlets) -> None:
+        self.idfElement = idfobj
+        self.name = idfobj.zone_name
+        self.type = objtype
+        self.air_inlets = inlets
+        self.air_outlets = outlets
+        self.data = {}
+        self.btype = None
+
+    def create(self):
+        self.btype = 'HVAC_Zone'
+        return self
